==> 02_scripts/09_pol_ideo_models.qmd <==
---
title: "Analysis of Motivated Reasoning on Political Topics Using Ideo Motives"
format:
  html:
    theme:
      light: lux
    toc: true
    toc-depth: 4
embed-resources: true
---

```{r setup, include=FALSE}
# knit settings
knitr::opts_chunk$set(fig.align = "center", fig.retina = 3,
                      fig.width = 6, fig.height = (6 * 0.618),
                      out.width = "90%", collapse = FALSE)
options(digits = 3, width = 90)

# set errors messages to English
Sys.setenv(LANG = "en")

# Make all the random draws reproducible
set.seed(42)

# Turn of scientific notation
options(scipen=999)
```

```{r libraries, warning=T, message=T, include=F}
# use groundhog to make code maximally reproducible
if (!require("groundhog", quietly = TRUE)) {
  install.packages("groundhog")
}
library("groundhog")

# use groundhog to install and load packages
pkgs <- c("here",         # Path management
          "tidyverse",    # ggplot, dplyr, %>%, and friends
          "ordbetareg"    # A brms wrapper to fit ordered beta regressions
          )

groundhog.library(pkgs, "2024-07-01") 
```

```{r Stan backend, include=FALSE}
# I use the faster cmdstanr backend for Stan in the place of rstan
# to do so install the cmdstanr package first: https://mc-stan.org/cmdstanr/
# Then run cmdstanr::install_cmdstan() to install cmdstan on your computer.
# RTools must be installed as well. For more info see 
# https://mc-stan.org/cmdstanr/articles/cmdstanr.html#introduction

# install.packages("cmdstanr", repos = c('https://stan-dev.r-universe.dev', getOption("repos")))
# cmdstanr::install_cmdstan()

options(mc.cores = 4,  
        brms.backend = "cmdstanr")
bayes_seed <- 42
```

```{r create-dirs, include=FALSE}
# set/create needed directories
model_dir <- here("03_output", "models", "pol_ideo")
dia_dir <- here("03_output", "diagnostics", "pol_ideo")

dir.create(model_dir, recursive = TRUE, showWarnings = TRUE)
dir.create(dia_dir, recursive = TRUE, showWarnings = TRUE)
```

## Data Preparations

#### Load data

```{r load-data, include=FALSE}
data_path <- here("01_data", "analysis", "data_analysis.RData")
load(file = data_path)
```

#### Filters

**Important: Filter out the responses to the Question (Screen != Question) as we only model the truth ratings for now!**

```{r pre-analysis-filter}
initial_rows <- nrow(data_analysis)
data_prep <- data_analysis %>% 
  filter(Screen == "Message")
filtered_rows <- initial_rows - nrow(data_prep)

filtered_rows
```

I also only want the political question topics and for these analyses I only keep Pro and Anti motive. 

```{r}
data_pol <- data_prep %>% 
  filter(question_type == "political") %>% 
  mutate(question_topic = factor(question_topic, 
                                 levels = c("climate", 
                                            "gender",         
                                            "immigration",
                                            "discrimination",
                                            "adoption",
                                            "punishment"))) %>%
  droplevels()   

levels(data_pol$question_topic)
```

#### Data types

As we will model monotonic effects, we have to transform the ideo_motive_strength variable into an ordered factor.

```{r ordered-factor-ideo-motive-strength}
data_pol <- data_pol %>%
  mutate(ideo_motive_strength = factor(ideo_motive_strength,
                                        levels = c("Anti-strong",
                                                   "Anti-moderate",
                                                   "Anti-weak",
                                                   "Neutral",
                                                   "Pro-weak",
                                                   "Pro-moderate",
                                                   "Pro-strong"),
                                        ordered = TRUE)) 
```

## Fit m1.pol

m1.pol: response_proportion \~ ideo_motive + (ideo_motive \| subj_idx) + (ideo_motive \| question_topic)

Check some settings:

```{r contrasts-m1.pol}
print(contrasts(data_pol$ideo_motive))
```

Set model formula

```{r formula-m1.pol}
f_m1.pol <- bf(response_proportion ~ ideo_motive + 
                 (ideo_motive | subj_idx) +
                 (ideo_motive | question_topic),  center = T)
```

And fit the ordered beta regression:

```{r fit-m1.pol, results='hide'}
m1.pol <- ordbetareg(formula = f_m1.pol,
                     data = data_pol, 
                     coef_prior_mean = 0,
                     coef_prior_SD = 1.5,
                     intercept_prior_mean = 0,
                     intercept_prior_SD = 1.5,
                     extra_prior = set_prior("lkj(2)", class = "L"),
                     seed = bayes_seed,
                     chains=4,
                     iter=4000, 
                     true_bounds = c(0, 1),
                     control = list(adapt_delta = 0.99,
                                    max_treedepth = 12),
                     file = here(model_dir, "m1.pol"))
```

#### Model checks

```{r print-m1.pol}
print(m1.pol)
```

##### Posterior predictive check

```{r m1.pol-ppc, message=FALSE, warning=FALSE, fig.height=6, fig.width=12}
pp_check_ordbeta(m1.pol, ndraws = 50)

ggsave(here(dia_dir, "ppc_m1.pol.png"), 
       width = 12, height = 6)
```

##### Model convergence

Check the trace and density plots.

```{r m1.pol-trace, fig.height=14, fig.width=18}
mcmc_plot(m1.pol, type = 'trace')
ggsave(here(dia_dir, "trace_plot_m1.pol.png"), 
       width = 18, height = 14)
```

```{r m1.pol-density, fig.height=14, fig.width=18}
mcmc_plot(m1.pol, type = 'dens')
ggsave(here(dia_dir, "density_plot_m1.pol.png"), width = 18, height = 14)
```

## Fit m2.pol

m2.pol: response_proportion \~ ideo_motive_strength + (ideo_motive_strength \| subj_idx) + (ideo_motive_strength \| question_topic)

Set model formula

```{r formula-m2.pol}
f_m2.pol <- bf(response_proportion ~ 
                 mo(ideo_motive_strength) + 
                 (mo(ideo_motive_strength) | subj_idx) +
                 (mo(ideo_motive_strength) | question_topic), center = T)
```

And fit the ordered beta regression:

```{r fit-m2.pol, results='hide'}
m2.pol <- ordbetareg(formula = f_m2.pol,
                     data = data_pol, 
                     coef_prior_mean = 0,
                     coef_prior_SD = 1.5,
                     intercept_prior_mean = 0,
                     intercept_prior_SD = 1.5,
                     extra_prior = set_prior("lkj(2)", class = "L"),
                     seed = bayes_seed,
                     chains=4,
                     iter=4000, 
                     true_bounds = c(0, 1),
                     control = list(adapt_delta = 0.99, 
                                    max_treedepth = 12),
                     file = here(model_dir, "m2.pol"))
```

#### Model checks

```{r print-m2.pol}
print(m2.pol)
```

##### Posterior predictive check

```{r m2.pol-ppc, message=FALSE, warning=FALSE, fig.height=6, fig.width=12}
pp_check_ordbeta(m2.pol, ndraws = 50)

ggsave(here(dia_dir, "ppc_m2.pol.png"), 
       width = 12, height = 6)
```

##### Model convergence

Check the trace and density plots:

```{r m2.pol-trace, fig.height=14, fig.width=18}
mcmc_plot(m2.pol, type = 'trace')
ggsave(here(dia_dir, "trace_plot_m2.pol.png"), 
       width = 18, height = 14)
```

```{r m2.pol-density, fig.height=14, fig.width=18}
mcmc_plot(m2.pol, type = 'dens')
ggsave(here(dia_dir, "density_plot_m2.pol.png"), width = 18, height = 14)
```

## Fit m3.pol

m3.pol: response_proportion \~ ideo_motive \* crt + (ideo_motive \| subj_idx) + (ideo_motive \| question_topic)

Set model formula

```{r formula-m3.pol}
f_m3.pol <- bf(response_proportion ~ ideo_motive * scale(crt_correct) + 
                 (ideo_motive | subj_idx) +
                 (ideo_motive | question_topic), center = T)
```

And fit the ordered beta regression:

```{r fit-m3.pol, results='hide'}
m3.pol <- ordbetareg(formula = f_m3.pol,
                     data = data_pol, 
                     coef_prior_mean = 0,
                     coef_prior_SD = 1.5,
                     intercept_prior_mean = 0,
                     intercept_prior_SD = 1.5,
                     extra_prior = set_prior("lkj(2)", class = "L"),
                     seed = bayes_seed,
                     chains=4,
                     iter=4000, 
                     true_bounds = c(0, 1),
                     control = list(adapt_delta = 0.99,                           
                                    max_treedepth = 12),
                     file = here(model_dir, "m3.pol"))
```

#### Model checks

```{r print-m3.pol}
print(m3.pol)
```

##### Posterior predictive check

```{r m3.pol-ppc, message=FALSE, warning=FALSE, fig.height=6, fig.width=12}
pp_check_ordbeta(m3.pol, ndraws = 50)

ggsave(here(dia_dir, "ppc_m3.pol.png"), 
       width = 12, height = 6)
```

##### Model convergence

Check the trace and density plots:

```{r m3.pol-trace, fig.height=14, fig.width=18}
mcmc_plot(m3.pol, type = 'trace')
ggsave(here(dia_dir, "trace_plot_m3.pol.png"), 
       width = 18, height = 14)
```

```{r m3.pol-density, fig.height=14, fig.width=18}
mcmc_plot(m3.pol, type = 'dens')
ggsave(here(dia_dir, "density_plot_m3.pol.png"), width = 18, height = 14)
```

## Fit m4.pol

m4.pol: response_proportion \~ ideo_motive \* commission_errors_r + (ideo_motive \| subj_idx) + (ideo_motive \| question_topic)

Set model formula

```{r formula-m4.pol}
f_m4.pol <- bf(response_proportion ~ ideo_motive * scale(commission_errors_r) + 
                 (ideo_motive | subj_idx) +
                 (ideo_motive | question_topic), center = T)
```

And fit the ordered beta regression:

```{r fit-m4.pol, results='hide'}
m4.pol <- ordbetareg(formula = f_m4.pol,
                     data = data_pol, 
                     coef_prior_mean = 0,
                     coef_prior_SD = 1.5,
                     intercept_prior_mean = 0,
                     intercept_prior_SD = 1.5,
                     extra_prior = set_prior("lkj(2)", class = "L"),
                     seed = bayes_seed,
                     chains=4,
                     iter=4000, 
                     true_bounds = c(0, 1),
                     control = list(adapt_delta = 0.99,
                                    max_treedepth = 12),
                     file = here(model_dir, "m4.pol"))
```

#### Model checks

```{r print-m4.pol}
print(m4.pol)
```

##### Posterior predictive check

```{r m4.pol-ppc, message=FALSE, warning=FALSE, fig.height=6, fig.width=12}
pp_check_ordbeta(m4.pol, ndraws = 50)

ggsave(here(dia_dir, "ppc_m4.pol.png"), 
       width = 12, height = 6)
```

##### Model convergence

Check the trace and density plots:

```{r m4.pol-trace, fig.height=14, fig.width=18}
mcmc_plot(m4.pol, type = 'trace')
ggsave(here(dia_dir, "trace_plot_m4.pol.png"), 
       width = 18, height = 14)
```

```{r m4.pol-density, fig.height=14, fig.width=18}
mcmc_plot(m4.pol, type = 'dens')
ggsave(here(dia_dir, "density_plot_m4.pol.png"), width = 18, height = 14)
```

## Fit m5.pol

m5.pol: response_proportion \~ ideo_motive_strength \* crt + (ideo_motive_strength \| subj_idx) + (ideo_motive_strength \| question_topic)

Set model formula

```{r formula-m5.pol}
f_m5.pol <- bf(response_proportion ~ mo(ideo_motive_strength) * scale(crt_correct) + 
                 (mo(ideo_motive_strength) | subj_idx) +
                 (mo(ideo_motive_strength) | question_topic), center = T)
```

And fit the ordered beta regression:

```{r fit-m5.pol, results='hide'}
m5.pol <- ordbetareg(formula = f_m5.pol,
                     data = data_pol, 
                     coef_prior_mean = 0,
                     coef_prior_SD = 1.5,
                     intercept_prior_mean = 0,
                     intercept_prior_SD = 1.5,
                     extra_prior = set_prior("lkj(2)", class = "L"),
                     seed = bayes_seed,
                     chains=4,
                     iter=4000, 
                     true_bounds = c(0, 1),
                     control = list(adapt_delta = 0.99,
                                    max_treedepth = 12),
                     file = here(model_dir, "m5.pol"))
```

#### Model checks

```{r print-m5.pol}
print(m5.pol)
```

##### Posterior predictive check

```{r m5.pol-ppc, message=FALSE, warning=FALSE, fig.height=6, fig.width=12}
pp_check_ordbeta(m5.pol, ndraws = 50)

ggsave(here(dia_dir, "ppc_m5.pol.png"), 
       width = 12, height = 6)
```

##### Model convergence

Check the trace and density plots:

```{r m5.pol-trace, fig.height=14, fig.width=18}
mcmc_plot(m5.pol, type = 'trace')
ggsave(here(dia_dir, "trace_plot_m5.pol.png"), 
       width = 18, height = 14)
```

```{r m5.pol-density, fig.height=14, fig.width=18}
mcmc_plot(m5.pol, type = 'dens')
ggsave(here(dia_dir, "density_plot_m5.pol.png"), width = 18, height = 14)
```

## Fit m6.pol

m6.pol: response_proportion \~ ideo_motive_strength \* commission_errors_r + (ideo_motive_strength \| subj_idx) + (ideo_motive_strength \| question_topic)

Set model formula

```{r formula-m6.pol}
f_m6.pol <- bf(response_proportion ~ mo(ideo_motive_strength) * scale(commission_errors_r) + 
                 (mo(ideo_motive_strength) | subj_idx) +
                 (mo(ideo_motive_strength) | question_topic), center = T)
```

And fit the ordered beta regression:

```{r fit-m6.pol, results='hide'}
m6.pol <- ordbetareg(formula = f_m6.pol,
                     data = data_pol, 
                     coef_prior_mean = 0,
                     coef_prior_SD = 1.5,
                     intercept_prior_mean = 0,
                     intercept_prior_SD = 1.5,
                     extra_prior = set_prior("lkj(2)", class = "L"),
                     seed = bayes_seed,
                     chains=4,
                     iter=4000, 
                     true_bounds = c(0, 1),
                     control = list(adapt_delta = 0.99,
                                    max_treedepth = 12),
                     file = here(model_dir, "m6.pol"))
```

#### Model checks

```{r print-m6.pol}
print(m6.pol)
```

##### Posterior predictive check

```{r m6.pol-ppc, message=FALSE, warning=FALSE, fig.height=6, fig.width=12}
pp_check_ordbeta(m5.pol, ndraws = 50)

ggsave(here(dia_dir, "ppc_m6.pol.png"), 
       width = 12, height = 6)
```

##### Model convergence

Check the trace and density plots:

```{r m6.pol-trace, fig.height=14, fig.width=18}
mcmc_plot(m6.pol, type = 'trace')
ggsave(here(dia_dir, "trace_plot_m6.pol.png"), 
       width = 18, height = 14)
```

```{r m6.pol-density, fig.height=14, fig.width=18}
mcmc_plot(m6.pol, type = 'dens')
ggsave(here(dia_dir, "density_plot_m6.pol.png"), width = 18, height = 14)
```
